// src/Library/Twitch/Schemas/ITwitchVideoSchema.ts
export default interface ITwitchVideoSchema {
    created_at: string;
    description: string;
    duration: string;
    id: string;
    language: string;
    published_at: string;
    thumbnail_url: string;
    title: string;
    type: string;
    url: string;
    user_id: string;
    user_name: string;
    view_count: number;
    viewable: string;
}